Revert "Upgrade graphql-java from 15.0 to 20.0. Includes breaking changes from the underlying library."
This reverts commit e5408eb22eafb7edd824dcad7b2d82e5b8842e23.
