+++ liftwizard-maven-build/liftwizard-dependencies/pom.xml
@@ -369,13 +369,7 @@
             <dependency>
                 <groupId>com.graphql-java</groupId>
                 <artifactId>graphql-java</artifactId>
-                <version>20.0</version>
-            </dependency>
-
-            <dependency>
-                <groupId>com.graphql-java</groupId>
-                <artifactId>graphql-java-extended-scalars</artifactId>
-                <version>20.0</version>
+                <version>15.0</version>
             </dependency>
 
             <dependency>
